+++ src/main/kotlin/de/craftan/engine/GameEvents.kt
@@ -1,6 +1,5 @@
 package de.craftan.engine
 
-import net.ormr.eventbus.CancellableEvent
 import net.ormr.eventbus.Event
 
 /**
