+++ src/main/kotlin/de/craftan/engine/gameflow/GameFlow.kt
@@ -1,15 +1,24 @@
 package de.craftan.engine.gameflow
 
 import de.craftan.engine.CraftanPlayer
+import de.craftan.engine.gameflow.flows.TurnFlow
 
 abstract class GameFlow {
 
+    /**
+     * The current round index
+     * 0 => Before first round
+     * 1 => First round
+     * 2 => Second round
+     * etc.
+     */
     var roundIndex = 0
 
     /**
      * The current round of the game
      */
-    abstract var round: GameRound
+    // Todo Why the fuck is this nullable
+    var round: GameRound? = null
 
     /**
      * All participating players
@@ -22,9 +31,9 @@ abstract class GameFlow {
     abstract fun nextRound()
 
     /**
-     * Sequence of players in the rounds
-     * e.g. player2 -> player1 -> player4 -> player3 -> loop
+     * Init the GameFlow
      */
-    lateinit var playerSequence: TurnSequence
+    abstract fun init()
 
+    fun turnFlow(): TurnFlow = round?.turnSequence[round!!.turnIndex]?.second!!
 }
