+++ src/main/kotlin/de/craftan/engine/gameflow/GameRound.kt
@@ -1,32 +1,26 @@
 package de.craftan.engine.gameflow
 
 import de.craftan.engine.CraftanGame
+import de.craftan.bridge.CraftanBridgePlayer
 import de.craftan.engine.CraftanPlayer
 import de.craftan.engine.gameflow.flows.TurnFlow
 
 /**
  * Models a round of a CraftanGame
  */
-interface GameRound {
+abstract class GameRound {
     /**
      * Abstract name to be rendered ingame
      */
-    val name: String
-
-    /**
-     * The number of round from the beginning of the game
-     */
-    val roundNumber: Int
+    abstract val name: String
     
-    var turnIndex: Int
+    var turnIndex: Int = 0
 
-    val turnSequence:List<Pair<CraftanPlayer, TurnFlow>>
+    abstract val turnSequence: MutableList<Pair<CraftanPlayer, TurnFlow>>
 
-    /**
-     * The game the round is inside of
-     */
-    val game: CraftanGame
+    fun currentTurn(): Pair<CraftanPlayer, TurnFlow> = turnSequence[turnIndex]
 
+    // TODO should listen to turn end event and then call this
     fun nextTurn() {
         turnIndex++
         if (turnIndex >= turnSequence.size) {
