+++ src/main/kotlin/de/craftan/engine/gameflow/TurnState.kt
@@ -1,6 +1,7 @@
 package de.craftan.engine.gameflow
 
-import de.craftan.engine.CraftanGameAction
+import de.craftan.engine.gameflow.action.CraftanActionData
+import de.craftan.engine.gameflow.action.CraftanGameAction
 
 /**
  * Models the current state a round is at
@@ -19,5 +20,5 @@ abstract class TurnState {
     /**
      * A list of all possible actions for a player
      */
-    abstract val actions: List<CraftanGameAction<*>>
+    abstract val actions: List<CraftanGameAction<out CraftanActionData>>
 }
