+++ src/main/kotlin/de/craftan/engine/gameflow/rounds/InitGameRound.kt
@@ -1,8 +1,8 @@
 package de.craftan.engine.gameflow.rounds
 
-import de.craftan.engine.CraftanGame
 import de.craftan.engine.CraftanPlayer
 import de.craftan.engine.gameflow.GameRound
+import de.craftan.engine.gameflow.flows.InitTurnFlow
 import de.craftan.engine.gameflow.flows.TurnFlow
 
 /**
@@ -10,9 +10,20 @@ import de.craftan.engine.gameflow.flows.TurnFlow
  * where the players place their first settlements and roads,
  */
 class InitGameRound(
-    override val game: CraftanGame,
-    override val name: String = "Init game",
-    override val roundNumber: Int = 0,
-    override var turnIndex: Int = 0,
-    override val turnSequence: List<Pair<CraftanPlayer, TurnFlow>>,
-) : GameRound
+    val players: List<CraftanPlayer>,
+) : GameRound() {
+    override val turnSequence: MutableList<Pair<CraftanPlayer, TurnFlow>> = mutableListOf()
+    override val name: String = "Init game"
+
+    init {
+        // Add the init states individually
+        // Place the first settlement
+        players.forEach {
+            turnSequence.add(Pair(it, InitTurnFlow()))
+        }
+        // Place the second settlement
+        players.reversed().forEach {
+            turnSequence.add(Pair(it, InitTurnFlow()))
+        }
+    }
+}
