+++ src/main/kotlin/de/craftan/engine/map/GameTile.kt
@@ -122,3 +122,16 @@ fun toGameTiles(tilesInfo: List<List<TileInfo>>): List<GameTile> {
 
     return gametiles
 }
+
+/**
+ * Converts a list of rows of TileInformation to a map from the coordinate to the corresponding GameTile.
+ * Earlier in the outer List means the row is more up in the board.
+ * Earlier in the inner List means more to the left on the game board
+ * @param tilesInfo list of rows of TileInformation
+ */
+fun toCoordinateToGameTileMap(tilesInfo: List<List<TileInfo>>): MutableMap<TileCoordinate, GameTile> {
+    val gameTiles = toGameTiles(tilesInfo)
+    val map: MutableMap<TileCoordinate, GameTile> = mutableMapOf()
+    gameTiles.forEach { map[it.coordinate] = it }
+    return map
+}
